Modifico diseño para parecerse al wireframe
