+++ my-app/src/Sidebar.css
@@ -64,3 +64,10 @@ Note: Beware of modifying this element as it can break the animations - you shou
 .bm-overlay {
   background: rgba(0, 0, 0, 0.3);
 }
+
+/* TODO mover a otro CSS */
+.card--line1 {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
