+++ my-app/src/Sidebar/LugarCard.js
@@ -3,7 +3,12 @@ import React from "react";
 export default function LugarCard(props) {
     return (
         <div className="card">
-            <p>{props.titulo}</p>
+            <div className="card--line1">
+                <h3>{props.titulo}</h3>
+                <p>{props.categoria}</p>
+            </div>
+            <p>{props.puntuacion}/5</p>
+            <hr />
         </div>
     )
 }
